server/requirements.txt
flask==3.1.0
gunicorn==23.0.0
line-bot-sdk==3.14.0
Pillow==11.1.0
requests==2.32.3
